:recycle: [reefactor] 스테이지 play view의 onImage 함수를 camera view로 이동
