+++ lib/ui/view/stage/ml_kit_camera_play_view.dart
@@ -10,25 +10,26 @@ import 'package:google_mlkit_pose_detection/google_mlkit_pose_detection.dart';
 import 'package:pocket_pose/config/app_color.dart';
 import 'package:pocket_pose/config/audio_player/audio_player_util.dart';
 import 'package:pocket_pose/config/ml_kit/custom_pose_painter.dart';
+import 'package:pocket_pose/data/entity/socket_request/send_skeleton_request.dart';
 import 'package:pocket_pose/data/remote/provider/socket_stage_provider_impl.dart';
 import 'package:pocket_pose/data/remote/provider/stage_provider_impl.dart';
 import 'package:pocket_pose/domain/entity/stage_player_list_item.dart';
+import 'package:pocket_pose/domain/entity/stage_skeleton_pose_landmark.dart';
 import 'package:pocket_pose/main.dart';
 import 'package:provider/provider.dart';
 
 class MlKitCameraPlayView extends StatefulWidget {
-  const MlKitCameraPlayView(
-      {Key? key,
-      required this.onImage,
-      this.initialDirection = CameraLensDirection.back,
-      required this.isPlayer})
-      : super(key: key);
-  // 이미지 받을 때마다 실행하는 함수
-  final Function(InputImage inputImage) onImage;
+  const MlKitCameraPlayView({
+    Key? key,
+    this.initialDirection = CameraLensDirection.back,
+    required this.isPlayer,
+    required this.playerNum,
+  }) : super(key: key);
   // 카메라 렌즈 방향 변수
   final CameraLensDirection initialDirection;
   // 플레이어인지 확인하는 변수
   final bool isPlayer;
+  final int playerNum;
 
   @override
   State<MlKitCameraPlayView> createState() => _MlKitCameraPlayViewState();
@@ -56,11 +57,13 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
   ];
   int _seconds = 5;
   Timer? _timer;
-  final skeletonColorList = [
-    AppColor.mintNeonColor,
-    AppColor.yellowNeonColor,
-    AppColor.greenNeonColor
-  ];
+  // ml kit 변수
+  bool _canProcess = true;
+  bool _isBusy = false;
+  int _frameNum = 0;
+  // 스켈레톤 추출 변수 선언(google_mlkit_pose_detection 라이브러리)
+  final PoseDetector _poseDetector =
+      PoseDetector(options: PoseDetectorOptions());
   AssetsAudioPlayer? _assetsAudioPlayer;
   late StageProviderImpl _stageProvider;
   late SocketStageProviderImpl _socketStageProvider;
@@ -106,8 +109,8 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
       }
     }
 
-    // 카메라 실행 가능하면 포즈 추출 시작
-    if (_cameraIndex != -1) {
+    // 플레이어이면서 카메라 실행 가능하면 포즈 추출 시작
+    if (widget.isPlayer && _cameraIndex != -1) {
       _startLiveFeed();
     }
 
@@ -181,10 +184,12 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
 
   @override
   void dispose() {
+    _stopTimer();
+    _canProcess = false;
+    _poseDetector.close();
     _assetsAudioPlayer = null;
     _assetsAudioPlayer?.dispose();
     _controller?.dispose;
-    _stopTimer();
 
     super.dispose();
   }
@@ -211,6 +216,48 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
         buildCountdownWidget()
       ],
     );
+  }
+
+  void _paintSkeleton() {
+    var player1 = context.select<SocketStageProviderImpl,
+        Map<PoseLandmarkType, PoseLandmark>?>((provider) => provider.player1);
+    var player0 = context.select<SocketStageProviderImpl,
+        Map<PoseLandmarkType, PoseLandmark>?>((provider) => provider.player0);
+    var player2 = context.select<SocketStageProviderImpl,
+        Map<PoseLandmarkType, PoseLandmark>?>((provider) => provider.player2);
+
+    if (player1 != null) {
+      CustomPosePainter painterLeft = CustomPosePainter(
+          [Pose(landmarks: player1)],
+          const Size(1280.0, 720.0),
+          InputImageRotation.rotation270deg,
+          AppColor.yellowNeonColor);
+      _customPaintLeft = CustomPaint(painter: painterLeft);
+    } else {
+      _customPaintLeft = null;
+    }
+
+    if (player0 != null) {
+      CustomPosePainter painterMid = CustomPosePainter(
+          [Pose(landmarks: player0)],
+          const Size(1280.0, 720.0),
+          InputImageRotation.rotation270deg,
+          AppColor.mintNeonColor);
+      _customPaintMid = CustomPaint(painter: painterMid);
+    } else {
+      _customPaintMid = null;
+    }
+
+    if (player2 != null) {
+      CustomPosePainter painterRignt = CustomPosePainter(
+          [Pose(landmarks: player2)],
+          const Size(1280.0, 720.0),
+          InputImageRotation.rotation270deg,
+          AppColor.greenNeonColor);
+      _customPaintRight = CustomPaint(painter: painterRignt);
+    } else {
+      _customPaintRight = null;
+    }
   }
 
   // 플레이 화면: 플레이어 3명 스켈레톤 보임
@@ -332,7 +379,7 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
     });
   }
 
-  // 카메라에서 실시간으로 받아온 이미치 처리: PoseDetectorView에서 받아온 함수인 onImage(스켈레톤 전송) 실행
+  // 카메라에서 실시간으로 받아온 이미치 처리
   Future _processCameraImage(CameraImage image) async {
     final WriteBuffer allBytes = WriteBuffer();
     for (final Plane plane in image.planes) {
@@ -372,49 +419,41 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
     final inputImage =
         InputImage.fromBytes(bytes: bytes, inputImageData: inputImageData);
 
-    // PoseDetectorView에서 받아온 함수인 onImage(이미지에 포즈가 추출되었으면 스켈레톤 그려주는 함수) 실행
-    widget.onImage(inputImage);
+    // 스켈레톤 전송
+    _sendSkeleton(inputImage);
   }
 
-  void _paintSkeleton() {
-    var player1 = context.select<SocketStageProviderImpl,
-        Map<PoseLandmarkType, PoseLandmark>?>((provider) => provider.player1);
-    var player0 = context.select<SocketStageProviderImpl,
-        Map<PoseLandmarkType, PoseLandmark>?>((provider) => provider.player0);
-    var player2 = context.select<SocketStageProviderImpl,
-        Map<PoseLandmarkType, PoseLandmark>?>((provider) => provider.player2);
+  // 카메라에서 실시간으로 받아온 이미지 처리: 스켈레톤 전송
+  Future<void> _sendSkeleton(InputImage inputImage) async {
+    if (!_canProcess) return;
+    if (_isBusy) return;
+    _isBusy = true;
 
-    if (player1 != null) {
-      CustomPosePainter painterLeft = CustomPosePainter(
-          [Pose(landmarks: player1)],
-          const Size(1280.0, 720.0),
-          InputImageRotation.rotation270deg,
-          AppColor.yellowNeonColor);
-      _customPaintLeft = CustomPaint(painter: painterLeft);
-    } else {
-      _customPaintLeft = null;
-    }
+    // poseDetector에서 추출된 포즈 가져오기
+    List<Pose> poses = await _poseDetector.processImage(inputImage);
+    // 소켓에 스켈레톤 전송
+    for (final pose in poses) {
+      Map<String, StageSkeletonPoseLandmark> resultMap = {};
 
-    if (player0 != null) {
-      CustomPosePainter painterMid = CustomPosePainter(
-          [Pose(landmarks: player0)],
-          const Size(1280.0, 720.0),
-          InputImageRotation.rotation270deg,
-          AppColor.mintNeonColor);
-      _customPaintMid = CustomPaint(painter: painterMid);
-    } else {
-      _customPaintMid = null;
-    }
+      pose.landmarks.forEach((key, value) {
+        var poseLandmark = StageSkeletonPoseLandmark(
+            type: value.type.index,
+            x: value.x,
+            y: value.y,
+            z: value.z,
+            likelihood: value.likelihood);
 
-    if (player2 != null) {
-      CustomPosePainter painterRignt = CustomPosePainter(
-          [Pose(landmarks: player2)],
-          const Size(1280.0, 720.0),
-          InputImageRotation.rotation270deg,
-          AppColor.greenNeonColor);
-      _customPaintRight = CustomPaint(painter: painterRignt);
-    } else {
-      _customPaintRight = null;
+        resultMap[key.index.toString()] = poseLandmark;
+      });
+
+      var resuest = SendSkeletonRequest(
+          playerNum: widget.playerNum,
+          frameNum: _frameNum,
+          skeleton: resultMap);
+      _socketStageProvider.sendPlaySkeleton(resuest);
     }
+    _frameNum++;
+
+    _isBusy = false;
   }
 }
